+++ src/sys/os/browser/web/document.rs
@@ -10,7 +10,7 @@ use devela::_js_extern;
 #[cfg(feature = "unsafe_ffi")]
 use devela::{_js_doc, _js_method_str_alloc, js_bool, js_int32, js_uint32};
 
-#[doc = crate::_tags!(web)]
+#[doc = crate::_tags!(web uid)]
 /// Handle to the brower's global [Document] associated APIs.
 #[doc = crate::_doc_meta!{location("sys/os/browser/web")}]
 ///