+++ docs/changelog.md
@@ -353,13 +353,15 @@
 
 #### sys::os::linux::io
 - new types: `LinuxTermios<Input|Output|Control|Local>Flags`, `LinuxTermiosCC`, `LinuxTermiosCharSize`.
+- update `Linux`:
+  - change the signatures of the methods: `sig_handler`, `sig_handler_info` to improve ergonomics.
 - update `LinuxTermios`:
   - new methods: `update_state`, `enable_event_mode`, `make_event`, `make_raw`, `make_cooked_reset`, `set_echo`, `set_canonical`, `set_signals`, `set_extensions`, `set_break_interrupt`, `set_input_cr_to_lf`, `set_software_flow`, `set_output_processing`, `set_read_min_timeout`.
   - rename methods: `get_state` to `read_state`, `save_state` to `write_state`, `get_winsize` to `read_window_size`, `disable_raw_mode` to `reset_cooked_mode`
 - make `LINUX_TERMIOS_*` type namespaces private.
 
 #### sys::os::linux::process
-- new types: `LinuxSignalactionFlags`, `LinuxSignals`.
+- new types: `LinuxSigactionFlags`, `LinuxSignal`, `LinuxSignalSet`.
 - make `LINUX_SIG*` type namespaces private.
 - fix signal restorer & `sys_getpid` in x86.